refactor: pull builder out
